# api/openapi/kumwe-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kumwe CMS API",
    "version": "1.0.0",
    "description": "Authenticated, explicitly site-bound content management, public discovery, and health endpoints for Kumwe CMS 2.0."
  },
...
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/users": {
      "get": {"operationId": "listUsers", "security": [{"bearerAuth": [], "siteContext": []}], "responses": {"200": {"description": "User collection"}}},
      "post": {"operationId": "createUser", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"201": {"description": "User created"}}}
    },
    "/api/v1/users/{id}": {
      "patch": {"operationId": "updateUser", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"description": "User updated"}}}
    },
    "/api/v1/roles": {
      "get": {"operationId": "listRoles", "security": [{"bearerAuth": [], "siteContext": []}], "responses": {"200": {"description": "Role collection"}}},
      "post": {"operationId": "createRole", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"201": {"description": "Role created"}}}
    },
    "/api/v1/users/{id}/roles/{roleId}": {
      "put": {"operationId": "assignRole", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"description": "Role assigned"}}},
      "delete": {"operationId": "revokeRole", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"description": "Role revoked"}}}
    },
    "/api/v1/roles/{id}/grants": {
      "post": {"operationId": "createGrant", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"201": {"description": "Capability granted"}}}
    },
    "/api/v1/grants/{grantId}": {
      "delete": {"operationId": "revokeGrant", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"description": "Grant revoked"}}}
    },
    "/api/v1/tokens": {
      "get": {"operationId": "listTokens", "security": [{"bearerAuth": [], "siteContext": []}], "responses": {"200": {"description": "Token metadata without secrets"}}},
      "post": {"operationId": "createToken", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"201": {"description": "Token created; secret returned once"}}}
    },
    "/api/v1/tokens/{tokenId}": {
      "delete": {"operationId": "revokeToken", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"204": {"description": "Token revoked"}}}
    },
    "/api/v1/tokens/{tokenId}/rotate": {
      "post": {"operationId": "rotateToken", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "responses": {"201": {"description": "Replacement token created; old token revoked atomically"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "Kumwe opaque access token"},
      "siteContext": {
        "type": "apiKey",
        "in": "header",
        "name": "Kumwe-Site",
        "description": "Canonical site identifier. It must exactly match the site bound into the bearer token; host and forwarding headers never select a site."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Unique retry key. A completed response is replayed for 24 hours; reuse with a different request is rejected.",
        "schema": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"}
      },
...
    },
...
  }
}